Split airspaceTasks into seprate files
