+++ backend/worker.py
@@ -17,7 +17,10 @@ app = Celery(
     broker='amqp://',
     backend='amqp',
     include=[
-        'tasks.airspaceTasks',
+        'tasks.airspace.airspaces',
+        'tasks.airspace.centers',
+        'tasks.airspace.navaids',
+        'tasks.airspace.towers',
         'tasks.atmosphereTasks'
     ]
 )
